python_main_script/requirements.txt
numpy>=1.21
Pillow>=9.0
PyYAML>=6.0
matplotlib>=3.5
scipy>=1.7
opencv-python>=4.5
pynput>=1.7
pygame>=2.1
pyserial>=3.5
